requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.0.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
